Fixed tests by adding casts from int to short

+++ src/test/java/edu/leti/diplomserver/OKSResultControllerIntegrationTests.java
@@ -31,10 +31,10 @@ public class OKSResultControllerIntegrationTests {
         mockMvc.perform(MockMvcRequestBuilders
                 .post("/oksResult")
                 .content(asJsonString(OKSResult.builder()
-                        .answer1(1).answer2(2).answer3(3)
-                        .answer4(4).answer5(5).answer6(1)
-                        .answer7(2).answer8(3).answer9(4)
-                        .answer10(5).answer11(1).answer12(2).build()))
+                        .answer1((short) 1).answer2((short) 2).answer3((short) 3)
+                        .answer4((short) 4).answer5((short) 5).answer6((short) 1)
+                        .answer7((short) 2).answer8((short) 3).answer9((short) 4)
+                        .answer10((short) 5).answer11((short) 1).answer12((short) 2).build()))
                 .contentType(MediaType.APPLICATION_JSON)
                 .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk());
